tests/requirements-test.txt
# Testing Framework
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.11.1
pytest-benchmark==4.0.0

# API Testing
httpx==0.24.1
requests==2.31.0
fastapi-testclient==0.3.0

# Database Testing
sqlalchemy==2.0.20
alembic==1.12.0
factory-boy==3.3.0
faker==19.6.2

# Performance Testing
locust==2.15.1
memory-profiler==0.60.0
psutil==5.9.5

# Mocking and Stubbing
responses==0.23.3
vcrpy==5.0.0
freezegun==1.2.2

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.5.0
bandit==1.7.5
safety==2.3.5

# Test Data Generation
hypothesis==6.82.0

# Reporting
pytest-html==3.2.0
pytest-json-report==1.5.0
allure-pytest==2.13.2

# Integration Testing
testcontainers==4.5.0
docker==6.1.3